fix : refactor profile_badge.dart to widget directory
